changed int to bigint, added parseBook method, fixed exception description

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `books` (
-                         id INT NOT NULL AUTO_INCREMENT,
+                         id BIGINT NOT NULL AUTO_INCREMENT,
                          title VARCHAR(255) NOT NULL,
                          price INT NOT NULL,
                          PRIMARY KEY (id)
